+++ pwhash-core/src/main/java/de/slevermann/pwhash/argon2/Argon2iStrategy.java
@@ -16,6 +16,7 @@
 package de.slevermann.pwhash.argon2;
 
 import de.mkammerer.argon2.Argon2Factory;
+import de.slevermann.pwhash.InvalidHashException;
 
 /**
  * A strategy implementation using argon2i for password hashing.
@@ -41,8 +42,27 @@ public class Argon2iStrategy extends Argon2Strategy {
      * @param saltLength  the length of the generated salt
      * @param hashLength  the output length for the hash
      */
-    public Argon2iStrategy(int memoryCost, int parallelism, int timeCost, int saltLength, int hashLength) {
+    private Argon2iStrategy(int memoryCost, int parallelism, int timeCost, int saltLength, int hashLength) {
         super(memoryCost, parallelism, timeCost);
         this.argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2i, saltLength, hashLength);
+    }
+
+    /**
+     * Create a customized argon2i instance
+     * <p>
+     * For all arguments, there are default values present in superclass to be used if not all values need to be customized
+     *
+     * @param memoryCost  the memory cost in kibibytes
+     * @param parallelism the amount of threads to use
+     * @param timeCost    the amount of iterations to use
+     * @param saltLength  the length of the generated salt
+     * @param hashLength  the output length for the hash
+     * @return the customized argon2i instance
+     * @throws InvalidHashException if any of the parameters are invalid
+     */
+    public static Argon2iStrategy getInstance(int memoryCost, int parallelism, int timeCost, int saltLength, int hashLength)
+            throws InvalidHashException {
+        verifyParameters(memoryCost, parallelism, timeCost, saltLength, hashLength);
+        return new Argon2iStrategy(memoryCost, parallelism, timeCost, saltLength, hashLength);
     }
 }