+++ pwhash-core/src/test/java/de/slevermann/pwhash/argon2/Argon2iTest.java
@@ -15,21 +15,54 @@
  */
 package de.slevermann.pwhash.argon2;
 
+import de.slevermann.pwhash.InvalidHashException;
+import org.testng.annotations.Test;
+
 public class Argon2iTest extends Argon2Test {
 
-    Argon2iTest() {
+    Argon2iTest() throws InvalidHashException {
         this.id = "argon2i";
         this.defaultStrategy = new Argon2iStrategy();
-        this.customStrategy = new Argon2iStrategy(CUSTOM_MEMORY_COST, CUSTOM_PARALLELISM, CUSTOM_TIME_COST,
+        this.customStrategy = Argon2iStrategy.getInstance(CUSTOM_MEMORY_COST, CUSTOM_PARALLELISM, CUSTOM_TIME_COST,
                 Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
-        this.customMStrategy = new Argon2iStrategy(CUSTOM_MEMORY_COST,
+        this.customMStrategy = Argon2iStrategy.getInstance(CUSTOM_MEMORY_COST,
                 Argon2Strategy.DEFAULT_PARALLELISM, Argon2Strategy.DEFAULT_TIME_COST,
                 Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
-        this.customPStrategy = new Argon2iStrategy(Argon2Strategy.DEFAULT_MEMORY_COST,
+        this.customPStrategy = Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST,
                 CUSTOM_PARALLELISM, Argon2Strategy.DEFAULT_TIME_COST,
                 Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
-        this.customTStrategy = new Argon2iStrategy(Argon2Strategy.DEFAULT_MEMORY_COST,
+        this.customTStrategy = Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST,
                 Argon2Strategy.DEFAULT_PARALLELISM, CUSTOM_TIME_COST,
                 Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
+    }
+
+    @Test(expectedExceptions = InvalidHashException.class)
+    public void badMemoryCost() throws InvalidHashException {
+        Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_PARALLELISM * 4, Argon2Strategy.DEFAULT_PARALLELISM, Argon2Strategy.DEFAULT_TIME_COST,
+                Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
+    }
+
+    @Test(expectedExceptions = InvalidHashException.class)
+    public void badParallelism() throws InvalidHashException {
+        Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST, 0, Argon2Strategy.DEFAULT_TIME_COST,
+                Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
+    }
+
+    @Test(expectedExceptions = InvalidHashException.class)
+    public void badTimeCost() throws InvalidHashException {
+        Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST, Argon2Strategy.DEFAULT_PARALLELISM, 0,
+                Argon2Strategy.DEFAULT_SALT_LENGTH, Argon2Strategy.DEFAULT_HASH_LENGTH);
+    }
+
+    @Test(expectedExceptions = InvalidHashException.class)
+    public void badSaltLength() throws InvalidHashException {
+        Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST, Argon2Strategy.DEFAULT_PARALLELISM, Argon2Strategy.DEFAULT_TIME_COST,
+                7, Argon2Strategy.DEFAULT_HASH_LENGTH);
+    }
+
+    @Test(expectedExceptions = InvalidHashException.class)
+    public void badHashLength() throws InvalidHashException {
+        Argon2iStrategy.getInstance(Argon2Strategy.DEFAULT_MEMORY_COST, Argon2Strategy.DEFAULT_PARALLELISM, Argon2Strategy.DEFAULT_TIME_COST,
+                Argon2Strategy.DEFAULT_SALT_LENGTH, 3);
     }
 }